AdminController -> BookController 로 / TestController - 화면단 500 에러시 새로고침 기준 (서버연결확인)

+++ src/main/java/org/example/bookstore/vo/BookVO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class BookVO {
-    private long bookId;
+    private int bookId;
     private String title;
     private String author;
     private String category;
