+++ src/main/java/org/example/bookstore/vo/MenuVO.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class MenuVO {
 
-  private long menuId;
+  private int menuId;
   private String name;
   private String path;
-  private long parentId;
-  private long depth;
+  private int parentId;
+  private int depth;
 
 }
